unbreak event stuff vis-a-vis RawEventParsing's use...
... and, in particular, put back what'd seeemed at first glance
like unwanted copy constructors but that in fact are a sort of
'promotion construcor', used in the case in which e.g. you've
already got a ZESpatialMove event and you want to use it to build
a set of ZESpatialHarden (& -Soften) events that depict button
state changes; those events should mirror all the state naturally
'shared' between the two event types (e.g. timestamp, position, ...)