(1) important zoft extension; (2) Renderable's over & up zofted ...
1. One of zofts' most important features is expressed by the
'BecomeLike()' method, which is a sort of moral identity assignment:
the zoft on which it's called becomes functionally indistinguishable
from the zoft that is the method's argument. However, this wouldn't
work with a direct instance of the Zoft class itself, which has no
guts. So a new guts miniclass ('LatchGuts'), defined right there
internal to the Zoft class, adds a data object of type T that is
the target of operator-equals (only when the guts is a LatchGuts;
otherwise the operator does what it did and directly assigns to
top level 'val')... and this latchval gets propogated to all subscribing
zofts during the LatchGuts's Inhale() invocation. This last provides
a possibly-desirable 'delay' property, incidentally.

2. the m_over and m_up fields in Renderable are now ZoftVects. An
update to RectangleRenderable's Draw() method provides a primitive
means for respecting the associated geometric meaning.