checkpoint in the wild cormorant hunt for massive memory inflation; try ...
... adjusting the loop count at line 26 in Ticato.cpp, which controls
the number of times a MattedVideoRenderable is instantiated and deleted
(before, on the last go-'round, it's actually used and not deleted). The
loop limit is set to unity (1) at the outset, which is identical with
'regular' behavior. On a Hades Canyon machine with 32GB of RAM, memory
blows out and the program crashes after about five minutes. It's more
like a minute if the loop limit is set to 2; and at 3 death comes within
seconds...